Change hashbang from sh to bash
The script is written using Bash syntax, as is written in the
project description at https://github.com/rauchg/wifi-password
"Get the password of the wifi you're on (bash)"

Using sh in the hashbang line makes this script crash on systems
that don't use bash to run sh script, but e.g. dash or some other
shell with no non-standard extensions to the POSIX shell spec.

Changing the hashbang to bash fixes that.